声明ServerModel的方法。调整ClientModel中的参数类型，包括返回值使用int代替原来的size_t，并约定一次报文最多2G，再多则需要增加续传逻辑。

+++ test/net_aio/net_aio.cpp
@@ -20,12 +20,12 @@ class FakeIoService
 {
 public:
     template<class Buf>
-    ssize_t PostRead(int channel, const Buf &buf, int timeout) {
+    int PostRead(int channel, const Buf &buf, int timeout) {
         return 0;
     }
 
     template<class Buf>
-    ssize_t PostWrite(int channel, const Buf &buf, int timeout) {
+    int PostWrite(int channel, const Buf &buf, int timeout) {
         return 0;
     }
 };
@@ -36,6 +36,7 @@ int test_client_model(int argc, char **argv)
     FakeChannelPool pool;
     FakeIoService   iosvc;
     en::ClientModel<FakeChannelPool, FakeIoService> client(pool, iosvc);
+    
     return 0;
 }
